Spi data transfer pointer is now the correct size
- As mentioned in the previous commit, the spi data register is a 32-bit
  register with only 16 accessible bits
- This now fixes any warning for incompatible pointer types when using
  those functions

+++ inc/spi.h
@@ -6,6 +6,6 @@
 void assert_spi_pin(volatile uint32_t* gpio_output_addr, unsigned int gpio_pin);
 void deassert_spi_pin(volatile uint32_t* gpio_output_addr, unsigned int gpio_pin);
 
-void trigger_spi_transfer(volatile uint16_t* spi_tx_reg, uint16_t data);
+void trigger_spi_transfer(volatile uint32_t* spi_tx_reg, uint16_t data);
 
 #endif /* SPI_H */
