+++ src/st7789.c
@@ -117,7 +117,7 @@ void st7789_hw_reset(struct St7789Internals* st7789_driver, void (*delay_us)(uns
 
 // Assumes no args for now
 void st7789_send_command(struct St7789Internals* st7789_driver
-                        , volatile uint16_t* spi_tx_reg
+                        , volatile uint32_t* spi_tx_reg
                         , uint8_t command_id)
 {
 	// DC/X is pulled lo to indicate a CMD being sent
